Revert "fix(slack): Fix Slack callback handling"
This reverts commit db018a94108553a7fb44225a97059ae968b31bca.
