+++ gradle.properties
@@ -8,7 +8,6 @@ testContainersVersion=1.15.3
 okHttpVersion=4.5.0
 resilience4jVersion=1.5.0
 spinnakerGradleVersion=8.23.0
-slackSdkVersion=1.12.1
 
 # Used to control whether to spin up docker to run liquibase before jooq
 buildingInDocker=false
