+++ keel-notifications/keel-notifications.gradle
@@ -6,9 +6,9 @@ dependencies {
   implementation(project(":keel-network"))
   implementation("org.springframework.boot:spring-boot-autoconfigure")
 
-  implementation("com.slack.api:slack-api-model-kotlin-extension:${slackSdkVersion}")
-  implementation("com.slack.api:slack-api-client-kotlin-extension:${slackSdkVersion}")
-  implementation("com.slack.api:bolt:${slackSdkVersion}")
+  implementation("com.slack.api:slack-api-model-kotlin-extension:1.4.1")
+  implementation("com.slack.api:slack-api-client-kotlin-extension:1.4.1")
+  implementation("com.slack.api:bolt:1.6.0")
 
   testImplementation("dev.minutest:minutest")
   testImplementation("io.strikt:strikt-core")
