--- keel-web/src/test/resources/slack/manual-judgement-payload.json
@@ -1,49 +1,0 @@
-{
-  "type": "block_actions",
-  "team": {
-    "id": "T9TK3CUKW",
-    "domain": "example"
-  },
-  "user": {
-    "id": "UA8RXUSPL",
-    "username": "jtorrance",
-    "team_id": "T9TK3CUKW"
-  },
-  "api_app_id": "AABA1ABCD",
-  "token": "9s8d9as89d8as9d8as989",
-  "container": {
-    "type": "message_attachment",
-    "message_ts": "1548261231.000200",
-    "attachment_id": 1,
-    "channel_id": "CBR2V3XEX",
-    "is_ephemeral": false,
-    "is_app_unfurl": false
-  },
-  "trigger_id": "12321423423.333649436676.d8c1bb837935619ccad0f624c448ffb3",
-  "channel": {
-    "id": "CBR2V3XEX",
-    "name": "review-updates"
-  },
-  "message": {
-    "bot_id": "BAH5CA16Z",
-    "type": "message",
-    "text": "Do you want to deploy version X into deployment Y of app A?",
-    "user": "UAJ2RU415",
-    "ts": "1548261231.000200"
-  },
-  "response_url": "https://hooks.slack.com/actions/AABA1ABCD/1232321423432/D09sSasdasdAS9091209",
-  "actions": [
-    {
-      "action_id": "constraintId:OVERRIDE_PASS:MANUAL_JUDGMENT",
-      "block_id": "=qXel",
-      "text": {
-        "type": "plain_text",
-        "text": "Manual judgement",
-        "emoji": true
-      },
-      "value": "approve",
-      "type": "button",
-      "action_ts": "1548426417.840180"
-    }
-  ]
-}
